Extracted codes for formatting into the class Formatter
