+++ src/seedu/addressbook/common/Messages.java
@@ -15,4 +15,5 @@ public class Messages {
             "java seedu.addressbook.Main [STORAGE_FILE_PATH]";
     public static final String MESSAGE_WELCOME = "Welcome to your Address Book!";
     public static final String MESSAGE_USING_STORAGE_FILE = "Using storage file : %1$s";
+    public static final String MESSAGE_ENTER_COMMAND = "Enter command: ";
 }
