+++ src/seedu/addressbook/ui/TextUi.java
@@ -1,16 +1,7 @@
 package seedu.addressbook.ui;
 
-import static seedu.addressbook.common.Messages.*;
-
-import seedu.addressbook.commands.CommandResult;
-import seedu.addressbook.common.Utils;
-import seedu.addressbook.data.person.ReadOnlyPerson;
-
 import java.io.InputStream;
 import java.io.PrintStream;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
 import java.util.Scanner;
 
 /**
@@ -18,152 +9,77 @@ import java.util.Scanner;
  */
 public class TextUi {
 
-    /** A decorative prefix added to the beginning of lines printed by AddressBook */
-    private static final String LINE_PREFIX = "|| ";
+	/**
+	 * Format of a comment input line. Comment lines are silently consumed when
+	 * reading user input.
+	 */
+	private static final String COMMENT_LINE_FORMAT_REGEX = "#.*";
 
-    /** A platform independent line separator. */
-    private static final String LS = System.lineSeparator();
+	private final Scanner in;
+	public final Formatter formatter;
 
-    private static final String DIVIDER = "===================================================";
+	public TextUi() {
+		this(System.in, System.out);
+	}
 
-    /** Format of indexed list item */
-    private static final String MESSAGE_INDEXED_LIST_ITEM = "\t%1$d. %2$s";
+	public TextUi(InputStream in, PrintStream out) {
+		this.in = new Scanner(in);
+		this.formatter = new Formatter(out);
+	}
 
+	/**
+	 * Returns true if the user input line should be ignored. Input should be
+	 * ignored if it is parsed as a comment, is only whitespace, or is empty.
+	 *
+	 * @param rawInputLine
+	 *            full raw user input line.
+	 * @return true if the entire user input line should be ignored.
+	 */
+	private boolean shouldIgnore(String rawInputLine) {
+		return rawInputLine.trim().isEmpty() || isCommentLine(rawInputLine);
+	}
 
-    /** Offset required to convert between 1-indexing and 0-indexing.  */
-    public static final int DISPLAYED_INDEX_OFFSET = 1;
+	/**
+	 * Returns true if the user input line is a comment line.
+	 *
+	 * @param rawInputLine
+	 *            full raw user input line.
+	 * @return true if input line is a comment.
+	 */
+	private boolean isCommentLine(String rawInputLine) {
+		return rawInputLine.trim().matches(COMMENT_LINE_FORMAT_REGEX);
+	}
 
-    /** Format of a comment input line. Comment lines are silently consumed when reading user input. */
-    private static final String COMMENT_LINE_FORMAT_REGEX = "#.*";
+	/**
+	 * Prompts for the command and reads the text entered by the user. Ignores
+	 * empty, pure whitespace, and comment lines. Echos the command back to the
+	 * user.
+	 * 
+	 * @return command (full line) entered by the user
+	 */
+	public String getUserCommand() {
+		formatter.showRequestUserCommand();
+		String fullInputLine = in.nextLine();
 
-    private final Scanner in;
-    private final PrintStream out;
+		// silently consume all ignored lines
+		while (shouldIgnore(fullInputLine)) {
+			fullInputLine = in.nextLine();
+		}
 
-    public TextUi(){
-        this(System.in, System.out);
-    }
+		formatter.showToUser("[Command entered:" + fullInputLine + "]");
+		return fullInputLine;
+	}
 
-    public TextUi(InputStream in, PrintStream out) {
-        this.in = new Scanner(in);
-        this.out = out;
-    }
+	public void showWelcomeMessage(String version, String storageFilePath) {
+		formatter.showWelcomeMessage(version, storageFilePath);
+	}
 
-    /**
-     * Returns true if the user input line should be ignored.
-     * Input should be ignored if it is parsed as a comment, is only whitespace, or is empty.
-     *
-     * @param rawInputLine full raw user input line.
-     * @return true if the entire user input line should be ignored.
-     */
-    private boolean shouldIgnore(String rawInputLine) {
-        return rawInputLine.trim().isEmpty() || isCommentLine(rawInputLine);
-    }
+	public void showGoodbyeMessage() {
+		formatter.showGoodbyeMessage();
+	}
 
-    /**
-     * Returns true if the user input line is a comment line.
-     *
-     * @param rawInputLine full raw user input line.
-     * @return true if input line is a comment.
-     */
-    private boolean isCommentLine(String rawInputLine) {
-        return rawInputLine.trim().matches(COMMENT_LINE_FORMAT_REGEX);
-    }
-
-    /**
-     * Prompts for the command and reads the text entered by the user.
-     * Ignores empty, pure whitespace, and comment lines.
-     * Echos the command back to the user.
-     * @return command (full line) entered by the user
-     */
-    public String getUserCommand() {
-        out.print(LINE_PREFIX + "Enter command: ");
-        String fullInputLine = in.nextLine();
-
-        // silently consume all ignored lines
-        while (shouldIgnore(fullInputLine)) {
-            fullInputLine = in.nextLine();
-        }
-
-        showToUser("[Command entered:" + fullInputLine + "]");
-        return fullInputLine;
-    }
-
-
-    public void showWelcomeMessage(String version, String storageFilePath) {
-        String storageFileInfo = String.format(MESSAGE_USING_STORAGE_FILE, storageFilePath);
-        showToUser(
-                DIVIDER,
-                DIVIDER,
-                MESSAGE_WELCOME,
-                version,
-                MESSAGE_PROGRAM_LAUNCH_ARGS_USAGE,
-                storageFileInfo,
-                DIVIDER);
-    }
-
-    public void showGoodbyeMessage() {
-        showToUser(MESSAGE_GOODBYE, DIVIDER, DIVIDER);
-    }
-
-
-    public void showInitFailedMessage() {
-        showToUser(MESSAGE_INIT_FAILED, DIVIDER, DIVIDER);
-    }
-
-    /** Shows message(s) to the user */
-    public void showToUser(String... message) {
-        for (String m : message) {
-            out.println(LINE_PREFIX + m.replace("\n", LS + LINE_PREFIX));
-        }
-    }
-
-    /**
-     * Shows the result of a command execution to the user. Includes additional formatting to demarcate different
-     * command execution segments.
-     */
-    public void showResultToUser(CommandResult result) {
-        final Optional<List<? extends ReadOnlyPerson>> resultPersons = result.getRelevantPersons();
-        if(resultPersons.isPresent()) {
-            showPersonListView(resultPersons.get());
-        }
-        showToUser(result.feedbackToUser, DIVIDER);
-    }
-
-    /**
-     * Shows a list of persons to the user, formatted as an indexed list.
-     * Private contact details are hidden.
-     */
-    private void showPersonListView(List<? extends ReadOnlyPerson> persons) {
-        final List<String> formattedPersons = new ArrayList<>();
-        for (ReadOnlyPerson person : persons) {
-            formattedPersons.add(person.getAsTextHidePrivate());
-        }
-        showToUserAsIndexedList(formattedPersons);
-    }
-
-    /** Shows a list of strings to the user, formatted as an indexed list. */
-    private void showToUserAsIndexedList(List<String> list) {
-        showToUser(getIndexedListForViewing(list));
-    }
-
-    /** Formats a list of strings as a viewable indexed list. */
-    private static String getIndexedListForViewing(List<String> listItems) {
-        final StringBuilder formatted = new StringBuilder();
-        int displayIndex = 0 + DISPLAYED_INDEX_OFFSET;
-        for (String listItem : listItems) {
-            formatted.append(getIndexedListItem(displayIndex, listItem)).append("\n");
-            displayIndex++;
-        }
-        return formatted.toString();
-    }
-
-    /**
-     * Formats a string as a viewable indexed list item.
-     *
-     * @param visibleIndex visible index for this listing
-     */
-    private static String getIndexedListItem(int visibleIndex, String listItem) {
-        return String.format(MESSAGE_INDEXED_LIST_ITEM, visibleIndex, listItem);
-    }
+	public void showInitFailedMessage() {
+		formatter.showInitFailedMessage();
+	}
 
 }
